+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/BilletRepository.java
@@ -4,5 +4,5 @@ import com.MIAGE.jeuxmiagiques.model.Billet;
 //JpaRepository is particularly a JPA specific extension for Repository. It has full API CrudRepository and PagingAndSortingRepository. So, basically, Jpa Repository contains the APIs for basic CRUD operations, the APIS for pagination, and the APIs for sorting.
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BilletRepository extends JpaRepository<Billet, Long> {
+public interface BilletRepository extends JpaRepository<Billet, Integer> {
 }
