+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/EpreuveRepository.java
@@ -3,5 +3,5 @@ package com.MIAGE.jeuxmiagiques.repository;
 import com.MIAGE.jeuxmiagiques.model.Epreuve;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EpreuveRepository extends JpaRepository<Epreuve, Long> {
+public interface EpreuveRepository extends JpaRepository<Epreuve, Integer> {
 }
