+++ src/main/java/com/MIAGE/jeuxmiagiques/repository/OrganisateurRepository.java
@@ -3,5 +3,5 @@ package com.MIAGE.jeuxmiagiques.repository;
 import com.MIAGE.jeuxmiagiques.model.Organisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganisateurRepository extends JpaRepository<Organisateur, Long> {
+public interface OrganisateurRepository extends JpaRepository<Organisateur, Integer> {
 }
